+++ packages/core/src/html/render.mjs
@@ -29,7 +29,6 @@
  */
 
 import fs from 'node:fs';
-import os from 'node:os';
 import path from 'node:path';
 import {
   CHROME,